Be lenient about call graph root
I have to admit that I'm not sure if a root that's not Node.ROOT breaks important assumptions that need to hold for the profile numbers to be correct, I'll hopefully find some time to look at that in detail. In the meantime, this fix allows me to see some numbers that seem to make sense.